+++ demo/package.json
@@ -71,7 +71,7 @@
     "react-hook-form": "^8.0.0-alpha.4",
     "react-redux": "^8.0.5",
     "react-router-bootstrap": "^0.26.2",
-    "react-router-dom": "^6.6.1",
+    "react-router-dom": "^6.6.2",
     "redux": "^4.2.0",
     "sass": "^1.57.1",
     "sequelize": "^6.28.0",
