+++ package.json
@@ -54,12 +54,12 @@
     "@grafana/eslint-config": "^5.1.0",
     "@rollup/plugin-commonjs": "^24.0.0",
     "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-terser": "^0.2.1",
-    "@rollup/plugin-typescript": "^10.0.1",
+    "@rollup/plugin-terser": "^0.3.0",
+    "@rollup/plugin-typescript": "^11.0.0",
     "@types/jest": "^29.2.5",
     "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.48.0",
-    "@typescript-eslint/parser": "^5.48.0",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
     "cypress": "^12.3.0",
     "eslint": "^8.31.0",
     "eslint-config-prettier": "^8.6.0",
@@ -67,7 +67,7 @@
     "eslint-plugin-jsdoc": "^39.6.4",
     "eslint-plugin-no-only-tests": "^3.1.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react": "^7.32.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^8.0.3",
     "jest": "^29.3.1",
@@ -76,10 +76,10 @@
     "lint-staged": "^13.1.0",
     "madge": "^5.0.1",
     "markdownlint": "^0.27.0",
-    "markdownlint-cli": "^0.32.2",
+    "markdownlint-cli": "^0.33.0",
     "npm-run-all": "^4.1.5",
     "os-browserify": "^0.3.0",
-    "prettier": "^2.8.1",
+    "prettier": "^2.8.2",
     "process": "^0.11.10",
     "rimraf": "^3.0.2",
     "rollup": "^3.9.1",
