+++ packages/core/src/extensions/index.ts
@@ -1,0 +1,3 @@
+export { BaseExtension } from './baseExtension';
+
+export type { Extension } from './types';
