+++ packages/core/src/instrumentations/base.ts
@@ -1,6 +1,6 @@
 import type { API } from '../api';
+import { BaseExtension } from '../extensions';
 import type { Transports } from '../transports';
-import { BaseExtension } from '../utils';
 import type { Instrumentation } from './types';
 
 export abstract class BaseInstrumentation extends BaseExtension implements Instrumentation {
