+++ packages/core/src/instrumentations/types.ts
@@ -1,6 +1,6 @@
 import type { API } from '../api';
+import type { Extension } from '../extensions';
 import type { Transports } from '../transports';
-import type { Extension } from '../utils';
 
 export interface Instrumentation extends Extension {
   api: API;
