+++ packages/core/src/transports/base.ts
@@ -1,4 +1,4 @@
-import { BaseExtension } from '../utils';
+import { BaseExtension } from '../extensions';
 import type { Transport, TransportItem } from './types';
 
 export abstract class BaseTransport extends BaseExtension implements Transport {
