+++ packages/core/src/transports/types.ts
@@ -1,7 +1,7 @@
 import type { APIEvent, EventEvent, ExceptionEvent, LogEvent, MeasurementEvent, TraceEvent } from '../api';
 import type { Patterns } from '../config';
+import type { Extension } from '../extensions';
 import type { Meta } from '../metas';
-import type { Extension } from '../utils';
 import type { TransportItemType } from './const';
 
 export type TransportItemPayload<P = APIEvent> = P;
