+++ packages/react/package.json
@@ -59,7 +59,7 @@
     "@types/hoist-non-react-statics": "^3.3.1",
     "@types/react": "^18.0.26",
     "react": "^18.2.0",
-    "react-router-dom": "^6.6.1"
+    "react-router-dom": "^6.6.2"
   },
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0 || ^18.0.0",
