+++ packages/react/src/index.ts
@@ -74,6 +74,7 @@ export {
   defaultViewMeta,
   ErrorsInstrumentation,
   FetchTransport,
+  genShortID,
   getCurrentTimestamp,
   getDataFromSafariExtensions,
   getInternalFaroFromGlobalObject,
@@ -148,6 +149,7 @@ export type {
   ExceptionsAPI,
   ExtendedError,
   ExtendedPromiseRejectionEvent,
+  Extension,
   FetchTransportOptions,
   FetchTransportRequestOptions,
   GlobalObject,
