+++ packages/web-sdk/package.json
@@ -53,7 +53,7 @@
   "dependencies": {
     "@grafana/faro-core": "^1.0.0-beta4",
     "ua-parser-js": "^1.0.32",
-    "web-vitals": "^3.1.0"
+    "web-vitals": "^3.1.1"
   },
   "devDependencies": {
     "@types/ua-parser-js": "^0.7.36"
