+++ packages/web-sdk/src/index.ts
@@ -41,6 +41,7 @@ export {
   defaultGlobalObjectKey,
   defaultInternalLoggerLevel,
   defaultLogLevel,
+  genShortID,
   getCurrentTimestamp,
   getInternalFaroFromGlobalObject,
   getTransportBody,
@@ -102,6 +103,7 @@ export type {
   ExceptionStackFrame,
   ExceptionsAPI,
   ExtendedError,
+  Extension,
   GlobalObject,
   Instrumentation,
   Instrumentations,
